Created Bozon
Bozon will be a store in which all applications are routed to. It creates a store object including the JSX component and other population req. this allows the reactor to communicate solely to it when the toolkit-state is in stack exa.This will also allow for easy authorizatrion, customization, meta types to be attactech to the objec in store before being renderd into a sub frame.